requirements.txt
alembic==1.18.4
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==3.2.2
billiard==4.2.4
boto3==1.42.55
botocore==1.42.55
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.5
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.131.0
fastapi-cli==0.0.23
fastapi-cloud-cli==0.13.0
fastar==0.8.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jmespath==1.1.0
kombu==5.6.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==26.0
passlib==1.7.4
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
redis==7.2.0
rich==14.3.3
rich-toolkit==0.19.4
rignore==0.7.6
rsa==4.9.1
s3transfer==0.16.0
sentry-sdk==2.53.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
